+++ src/main/java/kcs/funding/fundingboost/domain/controller/BookmarkController.java
@@ -1,0 +1,9 @@
+package kcs.funding.fundingboost.domain.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class BookmarkController {
+}
