+++ src/main/java/kcs/funding/fundingboost/domain/controller/FundingController.java
@@ -7,27 +7,38 @@ import kcs.funding.fundingboost.domain.dto.request.RegisterFundingDto;
 import kcs.funding.fundingboost.domain.dto.response.FriendFundingDetailDto;
 import kcs.funding.fundingboost.domain.dto.response.FriendFundingDto;
 import kcs.funding.fundingboost.domain.dto.response.FundingRegistrationItemDto;
+import kcs.funding.fundingboost.domain.dto.response.HomeViewDto;
+import kcs.funding.fundingboost.domain.dto.response.MyFundingHistoryDetailDto;
+import kcs.funding.fundingboost.domain.dto.response.MyFundingHistoryDto;
+import kcs.funding.fundingboost.domain.dto.response.MyFundingStatusDto;
+import kcs.funding.fundingboost.domain.dto.response.myPage.friendFundingHistory.FriendFundingHistoryDto;
 import kcs.funding.fundingboost.domain.service.FundingService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/v1/funding")
 @RequiredArgsConstructor
 public class FundingController {
 
     private final FundingService fundingService;
 
     /**
+     * 메인페이지 조회
+     */
+    @GetMapping("/api/v1/home")
+    public ResponseDto<HomeViewDto> home(@RequestParam("memberId") Long memberId) {
+        return ResponseDto.ok(fundingService.getMainView(memberId));
+    }
+
+    /**
      * 펀딩 등록 페이지 조회
      */
-    @GetMapping("")
+    @GetMapping("/api/v1/funding")
     public ResponseDto<List<FundingRegistrationItemDto>> viewFundingRegistration(
             @RequestParam(name = "memberId") Long memberId,
             @RequestParam(name = "ItemList") List<Long> registerFundingBringItemDto
@@ -38,7 +49,7 @@ public class FundingController {
     /**
      * 펀딩 등록하기
      */
-    @PostMapping("")
+    @PostMapping("/api/v1/funding")
     public ResponseDto<CommonSuccessDto> registerFunding(
             @RequestParam(name = "memberId") Long memberId,
             @RequestBody RegisterFundingDto registerFundingDto
@@ -50,7 +61,7 @@ public class FundingController {
     /**
      * 펀딩 종료하기
      */
-    @PostMapping("/close/{fundingId}")
+    @PostMapping("/api/v1/funding/close/{fundingId}")
     public ResponseDto<CommonSuccessDto> closeFunding(@PathVariable("fundingId") Long fundingId) {
         return ResponseDto.ok(fundingService.terminateFunding(fundingId));
     }
@@ -58,7 +69,7 @@ public class FundingController {
     /**
      * 친구 펀딩 디테일 페이지 조회
      */
-    @GetMapping("/friends/{fundingId}")
+    @GetMapping("/api/v1/funding/friends/{fundingId}")
     public ResponseDto<FriendFundingDetailDto> viewFriendsFundingDetail(@PathVariable("fundingId") Long fundingId,
                                                                         @RequestParam(name = "memberId") Long memberId) {
         return ResponseDto.ok(fundingService.viewFriendsFundingDetail(fundingId, memberId));
@@ -67,7 +78,7 @@ public class FundingController {
     /**
      * 친구 펀딩 목록 조회
      */
-    @GetMapping("/friends")
+    @GetMapping("/api/v1/funding/friends")
     public ResponseDto<List<FriendFundingDto>> viewFriendFundingList(
             @RequestParam(name = "memberId") Long memberId
     ) {
@@ -77,16 +88,42 @@ public class FundingController {
     /**
      * 펀딩 기간 늘리기
      */
-    @PostMapping("/extension/{fundingId}")
+    @PostMapping("/api/v1/funding/extension/{fundingId}")
     public ResponseDto<CommonSuccessDto> extendMyFunding(@PathVariable("fundingId") Long fundingId,
                                                          @RequestParam("memberId") Long memberId) {
         return ResponseDto.ok(fundingService.extendFunding(fundingId));
     }
-// 잔여금액 결제하기
-//    @PostMapping("/remain/{fundingItemId}")
-//    public ResponseDto<CommonSuccessDto> payRemain(@PathVariable("fundingItemId") Long fundingItemId,
-//                                                   @RequestBody PayRemainDto payRemainDto,
-//                                                   @RequestParam("memeberId") Long memberId) {
-//        return ResponseDto.ok(fundingService.payRemain(fundingItemId, payRemainDto, memberId));
-//    }
+
+    /**
+     * 마이 페이지 조회
+     */
+    @GetMapping("/api/v1/funding/my-funding-status")
+    public ResponseDto<MyFundingStatusDto> viewMyPage(@RequestParam("memberId") Long memberId) {
+        return ResponseDto.ok(fundingService.getMyFundingStatus(memberId));
+    }
+
+    /**
+     * 지난 펀딩 이력 조회
+     */
+    @GetMapping("/api/v1/funding/history")
+    public ResponseDto<MyFundingHistoryDto> viewMyFundingHistory(@RequestParam("memberId") Long memberId) {
+        return ResponseDto.ok(fundingService.getMyFundingHistory(memberId));
+    }
+
+    /**
+     * 지난 펀딩 이력 상세 조회
+     */
+    @GetMapping("/api/v1/funding/history/{fundingId}")
+    public ResponseDto<MyFundingHistoryDetailDto> viewMyFundingHistoryDetail(@RequestParam("memberId") Long memberId,
+                                                                             @PathVariable("fundingId") Long fundingId) {
+        return ResponseDto.ok(fundingService.getMyFundingHistoryDetails(memberId, fundingId));
+    }
+
+    /**
+     * 친구 펀딩 이력 조회
+     */
+    @GetMapping("/api/v1/funding/history/friend")
+    public ResponseDto<FriendFundingHistoryDto> viewFriendFundingHistory(@RequestParam("memberId") Long memberId) {
+        return ResponseDto.ok(fundingService.getFriendFundingHistory(memberId));
+    }
 }
