+++ src/main/java/kcs/funding/fundingboost/domain/service/ContributorService.java
@@ -1,0 +1,13 @@
+package kcs.funding.fundingboost.domain.service;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Slf4j
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+public class ContributorService {
+}
